backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
deepface==0.0.79
numpy==1.24.3
Pillow==10.1.0
gunicorn==21.2.0
tf-keras==2.16.0
